finish mixin refactor to base route class
