+++ ui/app/routes/vault/cluster/policies/policy.js
@@ -1,6 +1,6 @@
-import Route from '@ember/routing/route';
+import UnloadModelRecord from 'vault/routes/vault/unload-model-record';
 import { inject as service } from '@ember/service';
-export default class PolicyRoute extends Route {
+export default class PolicyRoute extends UnloadModelRecord {
   @service store;
 
   beforeModel() {
