--- ui/app/routes/vault/cluster/policies/policy/show.js
@@ -1,3 +1,0 @@
-import UnloadModelRecord from 'vault/routes/vault/cluster/unload-model-record';
-
-export default class PolicyShowRoute extends UnloadModelRecord {}
